g&s gestao, pequenos testes

+++ src/source/user.java
@@ -2,9 +2,9 @@ package source;
 
 public class user {
 	String nome, idade;
-	float salario;
+	double salario;
 
-	public user(String nome, String idade, float salario) {
+	public user(String nome, String idade, double salario) {
 		this.nome = nome;
 		this.idade = idade;
 		this.salario = salario;
@@ -26,7 +26,7 @@ public class user {
 		this.idade = idade;
 	}
 
-	public float getSalario() {
+	public double getSalario() {
 		return salario;
 	}
